Fix potential connection starvation in bmi_ib
The new coming connection are handled in the function
ib_tcp_server_check_new_connections(); and this is called inside the
function ib_block_for_activity(). However the ib_block_for_activity()
is only called when the network is idle in BMI_ib_testcontext() or
BMI_ib_testunexpected().

As a result, when the server is busy serving one client process, the
other processes can't make a new connections to the server and thus
they can't transfer data to the server concurrently.

Change-Id: Ib9c28e0e5c38cf03741c2220a9a95e2c9ffbe5a1